Refactor: Move math utilities from error_handling to math_utils
