+++ examples/trl_integration/dashboard_update_hooks_demo.py
@@ -15,7 +15,7 @@ import threading
 from pathlib import Path
 from typing import Dict, Any
 
-from rldk.utils.error_handling import safe_divide
+from rldk.utils.math_utils import safe_divide
 
 # Import TRL components
 try:
